requirements.txt
grpcio-tools ~= 1.75.1
botocore ~= 1.40.40
